# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API de Órdenes de Compra a Proveedores
  description: >
    Contrato de referencia para la implementación del backend. Este archivo es
    el contrato preciso contra el cual se implementa y se mantiene a mano.
  version: "2026-09-24"
  contact:
    name: Henry Molina
    url: https://henry-molina.dev/
servers:
  - url: http://localhost:8080/api
    description: Entorno local (Docker Compose)
security:
  - bearerAuth: []
...
paths:
...
  /webhooks/eventos:
    post:
      tags: [Webhooks]
      summary: Registrar un evento logístico reportado por el proveedor
      operationId: registrarEventoProveedor
      description: >
        Punto de integración para que el sistema del proveedor informe el
        avance de una orden (aceptada, despachada, entregada, etc.), sin
        depender de que un usuario interno la actualice a mano. Puramente
        informativo: ningún tipo de evento (incluido ENTREGADA o RECHAZADA)
        dispara una transición de estado interna de la orden. No usa el JWT
        del resto de la API: se autentica con la clave propia del proveedor, que
        identifica al proveedor (la URL no lleva su id). La orden se identifica por
        su numero_orden, el mismo que el proveedor recibe en el correo de aprobacion;
        una orden de otro proveedor responde 404 igual que una inexistente.
      security: [{ apiKeyAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [numero_orden, tipo_evento, fecha_evento]
              properties:
                numero_orden: { type: string, maxLength: 20 }
                tipo_evento: { $ref: "#/components/schemas/TipoEventoProveedor" }
                observacion: { type: string, maxLength: 250 }
                fecha_evento: { type: string, format: date-time }
      responses:
        "201":
          description: Evento registrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventoProveedor" }
        "401":
          description: API key ausente o inválida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: No existe una orden con ese numero_orden para este proveedor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT propio emitido por /auth/login (sin OAuth2)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Credencial por proveedor, usada exclusivamente por el endpoint de webhooks
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Clave generada por el cliente para evitar reprocesar la misma operación ante reintentos
...
  schemas:
...
    TipoEventoProveedor:
      type: string
      enum: [ACEPTADA, RECHAZADA, PREPARADA, DESPACHADA, ENTREGADA]
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [type, code, message]
          properties:
            type:
              type: string
              enum: [invalid_request_error, authentication_error, state_conflict_error, api_error]
            code: { type: string, example: categoria_no_permitida }
            message: { type: string }
            param: { type: string, nullable: true }
...
    EventoProveedor:
      type: object
      properties:
        object: { type: string, enum: [evento_proveedor] }
        id: { type: integer }
        orden_compra_id: { type: integer }
        tipo_evento: { $ref: "#/components/schemas/TipoEventoProveedor" }
        observacion: { type: string, nullable: true }
        fecha_evento: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...